moved getEscrowDeploymentInformation function to escrowInfoRetriever
