+++ packages/payment-detection/src/erc20/fee-proxy-contract.ts
@@ -19,18 +19,13 @@ const PROXY_CONTRACT_ADDRESS_MAP = {
   ['0.2.0']: '0.2.0',
 };
 
-const ESCROW_CONTRACT_ADDRESS_MAP = {
-  ['0.1.0']: '0.1.0',
-  ['0.2.0']: '0.1.0',
-};
-
 /**
  * Handle payment networks with ERC20 fee proxy contract extension, or derived
  */
 
 export abstract class ERC20FeeProxyPaymentDetectorBase<
   TExtension extends ExtensionTypes.PnFeeReferenceBased.IFeeReferenceBased,
-  TPaymentEventParameters extends PaymentTypes.IERC20FeePaymentEventParameters,
+  TPaymentEventParameters extends PaymentTypes.IERC20FeePaymentEventParameters
 > extends FeeReferenceBasedDetector<TExtension, TPaymentEventParameters> {
   /**
    * @param extension The advanced logic payment network extensions
@@ -103,8 +98,10 @@ export class ERC20FeeProxyPaymentDetector extends ERC20FeeProxyPaymentDetectorBa
       });
     }
 
-    const { address: proxyContractAddress, creationBlockNumber: proxyCreationBlockNumber } =
-      ERC20FeeProxyPaymentDetector.getDeploymentInformation(paymentChain, paymentNetwork.version);
+    const {
+      address: proxyContractAddress,
+      creationBlockNumber: proxyCreationBlockNumber,
+    } = ERC20FeeProxyPaymentDetector.getDeploymentInformation(paymentChain, paymentNetwork.version);
 
     if (networkSupportsTheGraph(paymentChain)) {
       const graphInfoRetriever = new TheGraphInfoRetriever(
@@ -139,10 +136,5 @@ export class ERC20FeeProxyPaymentDetector extends ERC20FeeProxyPaymentDetectorBa
   public static getDeploymentInformation = makeGetDeploymentInformation(
     erc20FeeProxyArtifact,
     PROXY_CONTRACT_ADDRESS_MAP,
-  );
-
-  public static getEscrowDeploymentInformation = makeGetDeploymentInformation(
-    erc20EscrowToPayArtifact,
-    ESCROW_CONTRACT_ADDRESS_MAP,
   );
 }
